# Integrate core search in multi-waypoint search and map matching