As super simple test using [this route](http://map.project-osrm.org/?z=14&center=52.524969,13.413019&loc=52.517037,13.388860&loc=52.514654,13.389587&loc=52.514941,13.392763&loc=52.513237,13.393450&loc=52.511311,13.393707&loc=52.508542,13.401260&loc=52.508529,13.402880&loc=52.508170,13.405219&loc=52.509117,13.406582&loc=52.509796,13.411045&loc=52.508020,13.414221&loc=52.509594,13.415401&loc=52.510906,13.416892&loc=52.511624,13.421774&loc=52.515020,13.426409&loc=52.516064,13.432417&loc=52.518206,13.429756&loc=52.520817,13.429756&loc=52.525569,13.429198&loc=52.526770,13.426838&hl=en&ly=&alt=&df=&srv=) (20 coordinates) on a core factor of `0.8`.

| Service | k=1.0 | k=0.8 (before) | k=0.8 (after) | speedup |
|------------|---------|----------------------|------------------|--------------|
| match | 56 ms | 645ms | 438 ms | x1.47 |
| viaroute | 4.21 ms | 96ms | 0.6 ms | x160 |

The viaroute probably hits an effect were contraction is bad for very local searches. For [this](http://map.project-osrm.org/?z=10&center=52.581774,13.705444&loc=52.565499,13.510437&loc=52.512460,13.428040&loc=52.475357,13.306847&hl=en&ly=&alt=&df=&srv=) request I see the following behaviour:

| Service | k=1.0 | k=0.8 (before) | k=0.8 (after) | speedup |
|------------|---------|----------------------|------------------|--------------|
| viaroute | 1.1ms | 68ms | 9.6 ms | x7.3 |