@TheMarex 2016-03-07
I slightly adapted @MoKob pruning method to generate the actual pruning value in a more robust fashion using pruning thresholds that we already utilized in `match`. The results are pretty great using the trace from above I'm seeing:

| Service | k=1.0 | k=0.8 (before) | k=0.8 (after) | speedup |
|------------|---------|----------------------|------------------|--------------|
| match | 56 ms | 438 ms | 58ms | x7.55 |

(note: The speedup for core factor 1.0 didn't change, apparently the pruning mostly hits the core edges)